+++ tests/Stub/fixtures/internal/expected.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Castor\Console;
+
+/** @internal */
+final class Application
+{
+    public function doStuff(): void
+    {
+    }
+}