+++ tests/Stub/fixtures/internal/input2.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Castor\Console;
+
+/** @internal */
+final class Application
+{
+    public function doStuff(): void
+    {
+        // some code
+    }
+}