+++ tests/Stub/fixtures/phpdoc-not-fqcn/expected.php
@@ -1,11 +1,13 @@
 <?php
 
-namespace Castor\Attribute;
+namespace Test\Attribute;
 
-class AsArgument extends \Castor\Attribute\AsCommandArgument
+use Castor\Attribute\AsCommandArgument;
+use Symfony\Component\Console\Completion\CompletionInput;
+class AsArgument extends AsCommandArgument
 {
     /**
-     * @param array<string>|callable(\Symfony\Component\Console\Completion\CompletionInput): array<string> $autocomplete
+     * @param array<string>|callable(CompletionInput): array<string> $autocomplete
      */
     public function test($autocomplete): void
     {