+++ tests/Stub/fixtures/same-namespace/expected.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Test\Foo;
+
+use Castor\Console\Application;
+use Castor\Kernel;
+function foo2(Application $application, Kernel $kernel): void
+{
+}
+function foo3(Application $application): void
+{
+}